# cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "這是後端 API 文件",
        "title": "interchange API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/proxy/{version}/{provider}/audio/speech": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "將文字透過指定 provider 的語音模型轉為音訊檔，回傳 `audio/*` 位元流。",
                "consumes": [
                    "application/json",
                    "multipart/form-data"
                ],
                "produces": [
                    "audio/*",
                    "application/octet-stream"
                ],
                "tags": [
                    "Proxy-Audio"
                ],
                "summary": "語音合成（Text-to-Speech）",
                "parameters": [
                    {
                        "enum": [
                            "v1"
                        ],
                        "type": "string",
                        "description": "API 版本",
                        "name": "version",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "openai"
                        ],
                        "type": "string",
                        "description": "提供者",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "語音合成請求內容",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_service_audio.AudioSpeechRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "音訊位元流（Content-Type 依實際模型回傳，例如 audio/mpeg、audio/wav）",
                        "schema": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_pkg_error.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_pkg_error.Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_pkg_error.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_pkg_error.Error"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_pkg_error.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_pkg_error.Error"
                        }
                    }
                }
            }
        },
        "/proxy/{version}/{provider}/audio/transcriptions": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "上傳音訊檔轉為文字（支援常見格式，如 OGG/MP3/WAV 等），回傳文字結果。",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Proxy-Audio"
                ],
                "summary": "語音轉錄（Speech-to-Text）",
                "parameters": [
                    {
                        "enum": [
                            "v1"
                        ],
                        "type": "string",
                        "description": "API 版本",
                        "name": "version",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "openai",
                            "gemini"
                        ],
                        "type": "string",
                        "description": "提供者",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "轉錄請求內容（含檔案與參數）",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_service_audio.AudioTranscriptionRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "轉錄文字內容",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/interchange_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_pkg_error.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_pkg_error.Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_pkg_error.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_pkg_error.Error"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_pkg_error.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_pkg_error.Error"
                        }
                    }
                }
            }
        },
        "/proxy/{version}/{provider}/audio/translations": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "上傳音訊檔並翻譯為指定語言的文字，回傳翻譯結果。",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Proxy-Audio"
                ],
                "summary": "語音翻譯（Speech Translation）",
                "parameters": [
                    {
                        "enum": [
                            "v1"
                        ],
                        "type": "string",
                        "description": "API 版本",
                        "name": "version",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "openai",
                            "gemini"
                        ],
                        "type": "string",
                        "description": "提供者",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "翻譯請求內容（含檔案與參數）",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_service_audio.AudioTranslationRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "翻譯後文字內容",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/interchange_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_pkg_error.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_pkg_error.Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_pkg_error.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_pkg_error.Error"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_pkg_error.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_pkg_error.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "interchange_internal_pkg_error.Error": {
            "type": "object"
        },
        "interchange_internal_pkg_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "description": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "requestID": {
                    "type": "string"
                }
            }
        },
        "interchange_internal_service_audio.AudioSpeechRequestBody": {
            "type": "object",
            "required": [
                "input",
                "model",
                "voice"
            ],
            "properties": {
                "input": {
                    "type": "string"
                },
                "instructions": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "response_format": {
                    "description": "mp3, wav, etc.",
                    "type": "string"
                },
                "speed": {
                    "description": "0.25 - 4.0",
                    "type": "number"
                },
                "stream_format": {
                    "description": "audio or sse",
                    "type": "string"
                },
                "voice": {
                    "type": "string"
                }
            }
        },
        "interchange_internal_service_audio.AudioTranscriptionRequestBody": {
            "type": "object",
            "required": [
                "file",
                "model"
            ],
            "properties": {
                "chunking_strategy": {
                    "description": "\"auto\" or blank",
                    "type": "string"
                },
                "file": {
                    "$ref": "#/definitions/multipart.FileHeader"
                },
                "include": {
                    "description": "optional logprobs",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "language": {
                    "description": "ISO-639-1 e.g. \"en\"",
                    "type": "string"
                },
                "model": {
                    "description": "whisper-1, gpt-4o-transcribe, etc.",
                    "type": "string"
                },
                "prompt": {
                    "type": "string"
                },
                "response_format": {
                    "description": "json, text, srt, etc.",
                    "type": "string"
                },
                "stream": {
                    "description": "gpt-4o only",
                    "type": "boolean"
                },
                "temperature": {
                    "type": "number"
                },
                "timestamp_granularities": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "interchange_internal_service_audio.AudioTranslationRequestBody": {
            "type": "object",
            "required": [
                "file",
                "model"
            ],
            "properties": {
                "file": {
                    "description": "使用 multipart/form-data 上傳檔案",
                    "allOf": [
                        {
                            "$ref": "#/definitions/multipart.FileHeader"
                        }
                    ]
                },
                "model": {
                    "description": "ex: whisper-1",
                    "type": "string"
                },
                "prompt": {
                    "type": "string"
                },
                "response_format": {
                    "description": "json, text, srt, verbose_json, vtt",
                    "type": "string"
                },
                "temperature": {
                    "description": "0 ~ 1",
                    "type": "number"
                }
            }
        },
...
        "multipart.FileHeader": {
            "type": "object",
            "properties": {
                "filename": {
                    "type": "string"
                },
                "header": {
                    "$ref": "#/definitions/textproto.MIMEHeader"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "textproto.MIMEHeader": {
            "type": "object",
            "additionalProperties": {
                "type": "array",
                "items": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "請在欄位輸入 \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}